// src/frontend/webapp/modilist/src/services/swagger/api/models/create-new-payment-method.ts
/* tslint:disable */
/* eslint-disable */
/**
 * ModilistAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface CreateNewPaymentMethod
 */
export interface CreateNewPaymentMethod {
    /**
     * 
     * @type {string}
     * @memberof CreateNewPaymentMethod
     */
    cardName?: string;
    /**
     * 
     * @type {string}
     * @memberof CreateNewPaymentMethod
     */
    cardNumber?: string;
    /**
     * 
     * @type {string}
     * @memberof CreateNewPaymentMethod
     */
    cardHolderName?: string;
    /**
     * 
     * @type {string}
     * @memberof CreateNewPaymentMethod
     */
    expireMonth?: string;
    /**
     * 
     * @type {string}
     * @memberof CreateNewPaymentMethod
     */
    expireYear?: string;
    /**
     * 
     * @type {boolean}
     * @memberof CreateNewPaymentMethod
     */
    isDefault?: boolean;
}